backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiosqlite>=0.22.1",
    "bcrypt==4.0.1",
    "fastapi>=0.129.2",
    "joblib>=1.5.3",
    "pandas>=3.0.1",
    "passlib[bcrypt]==1.7.4",
    "pydantic-settings>=2.13.1",
    "python-jose[cryptography]>=3.5.0",
    "python-multipart>=0.0.22",
    "scikit-learn>=1.8.0",
    "shap>=0.49.1",
    "sqlalchemy>=2.0.46",
    "uvicorn[standard]>=0.41.0",
    "xgboost>=3.2.0",
]

[dependency-groups]
dev = [
    "requests>=2.32.5",
]
